Move into tools/lint our clang-format-includes
