+++ tools/lint/BUILD.bazel
@@ -38,16 +38,23 @@ drake_py_binary(
     ],
 )
 
+# Depend on this when you need to call clang-format.  (We don't have a label
+# for the binary yet, but this will stand in for it in the meantime.)
+filegroup(
+    name = "clang-format",
+    srcs = ["//:.clang-format"],
+)
+
 drake_py_binary(
     name = "drakelint",
     srcs = ["drakelint.py"],
     # The clang-format-includes binary is not used by drakelint, because
     # drakelint imports the formatter library directly.  However, we add
     # clang-format-includes here as a data dependency because the formatter
-    # asks the user to run bazel-bin/drake/tools/clang-format-includes for
+    # asks the user to run bazel-bin/tools/lint/clang-format-includes for
     # fixups, and this ensures it always is freshly compiled and available.
-    data = ["//drake/tools:clang-format-includes"],
-    deps = ["//drake/tools:formatter"],
+    data = [":clang-format-includes"],
+    deps = [":formatter"],
 )
 
 drake_py_binary(
@@ -55,6 +62,20 @@ drake_py_binary(
     srcs = ["bzlcodestyle.py"],
     main = "bzlcodestyle.py",
     deps = ["@pycodestyle//:pycodestyle"],
+)
+
+drake_py_library(
+    name = "formatter",
+    srcs = ["formatter.py"],
+    data = [":clang-format"],
+)
+
+drake_py_binary(
+    name = "clang-format-includes",
+    srcs = ["clang_format_includes.py"],
+    data = [":clang-format"],
+    main = "clang_format_includes.py",
+    deps = [":formatter"],
 )
 
 # === test ===
@@ -75,6 +96,20 @@ drake_py_test(
         "local",
     ],
     deps = [":util"],
+)
+
+drake_py_test(
+    name = "formatter_test",
+    srcs = ["test/formatter_test.py"],
+    deps = [":formatter"],
+)
+
+sh_test(
+    name = "clang_format_includes_test",
+    srcs = ["test/clang_format_includes_test.sh"],
+    data = [
+        ":clang-format-includes",
+    ],
 )
 
 add_lint_tests()
